+++ src/QueryManager.ts
@@ -51,6 +51,7 @@ import {
 
 import {
   diffSelectionSetAgainstStore,
+  ResultTransformer,
 } from './data/diffAgainstStore';
 
 import {
@@ -191,6 +192,7 @@ export class QueryManager {
   private store: ApolloStore;
   private reduxRootKey: string;
   private queryTransformer: QueryTransformer;
+  private resultTransformer: ResultTransformer;
   private queryListeners: { [queryId: string]: QueryListener };
 
   private idCounter = 0;
@@ -222,6 +224,7 @@ export class QueryManager {
     store,
     reduxRootKey,
     queryTransformer,
+    resultTransformer,
     shouldBatch = false,
     batchInterval = 10,
   }: {
@@ -229,6 +232,7 @@ export class QueryManager {
     store: ApolloStore,
     reduxRootKey: string,
     queryTransformer?: QueryTransformer,
+    resultTransformer?: ResultTransformer,
     shouldBatch?: Boolean,
     batchInterval?: number,
   }) {
@@ -238,6 +242,7 @@ export class QueryManager {
     this.store = store;
     this.reduxRootKey = reduxRootKey;
     this.queryTransformer = queryTransformer;
+    this.resultTransformer = resultTransformer;
     this.pollingTimers = {};
     this.batchInterval = batchInterval;
     this.queryListeners = {};
@@ -379,12 +384,15 @@ export class QueryManager {
           }
         } else {
           const resultFromStore = readSelectionSetFromStore({
-            store: this.getDataWithOptimisticResults(),
+            context: {
+              store: this.getDataWithOptimisticResults(),
+              fragmentMap: queryStoreValue.fragmentMap,
+              resultTransformer: this.resultTransformer,
+            },
             rootId: queryStoreValue.query.id,
             selectionSet: queryStoreValue.query.selectionSet,
             variables: queryStoreValue.variables,
             returnPartialData: options.returnPartialData || options.noFetch,
-            fragmentMap: queryStoreValue.fragmentMap,
           });
 
           if (observer.next) {
@@ -622,12 +630,15 @@ export class QueryManager {
       // query, use the query diff algorithm to get as much of a result as we can, and identify
       // what data is missing from the store
       const { missingSelectionSets, result } = diffSelectionSetAgainstStore({
+        context: {
+          store: this.store.getState()[this.reduxRootKey].data,
+          fragmentMap: queryFragmentMap,
+          resultTransformer: this.resultTransformer,
+        },
         selectionSet: querySS.selectionSet,
-        store: this.store.getState()[this.reduxRootKey].data,
         throwOnMissingField: false,
         rootId: querySS.id,
         variables,
-        fragmentMap: queryFragmentMap,
       });
 
       initialResult = result;
@@ -727,12 +738,15 @@ export class QueryManager {
               // this will throw an error if there are missing fields in
               // the results if returnPartialData is false.
               resultFromStore = readSelectionSetFromStore({
-                store: this.getApolloState().data,
+                context: {
+                  store: this.getApolloState().data,
+                  fragmentMap: queryFragmentMap,
+                  resultTransformer: this.resultTransformer,
+                },
                 rootId: querySS.id,
                 selectionSet: querySS.selectionSet,
                 variables,
                 returnPartialData: returnPartialData || noFetch,
-                fragmentMap: queryFragmentMap,
               });
               // ensure multiple errors don't get thrown
               /* tslint:disable */