+++ src/data/readFromStore.ts
@@ -1,5 +1,6 @@
 import {
   diffSelectionSetAgainstStore,
+  StoreContext,
 } from './diffAgainstStore';
 
 import {
@@ -10,7 +11,6 @@ import {
 import {
   getQueryDefinition,
   getFragmentDefinition,
-  FragmentMap,
 } from '../queries/getFromAST';
 
 import {
@@ -35,7 +35,7 @@ export function readQueryFromStore({
   const queryDef = getQueryDefinition(query);
 
   return readSelectionSetFromStore({
-    store,
+    context: { store, fragmentMap: {} },
     rootId: 'ROOT_QUERY',
     selectionSet: queryDef.selectionSet,
     variables,
@@ -59,7 +59,7 @@ export function readFragmentFromStore({
   const fragmentDef = getFragmentDefinition(fragment);
 
   return readSelectionSetFromStore({
-    store,
+    context: { store, fragmentMap: {} },
     rootId,
     selectionSet: fragmentDef.selectionSet,
     variables,
@@ -68,29 +68,26 @@ export function readFragmentFromStore({
 }
 
 export function readSelectionSetFromStore({
-  store,
+  context,
   rootId,
   selectionSet,
   variables,
   returnPartialData = false,
-  fragmentMap,
 }: {
-  store: NormalizedCache,
+  context: StoreContext,
   rootId: string,
   selectionSet: SelectionSet,
   variables: Object,
   returnPartialData?: boolean,
-  fragmentMap?: FragmentMap,
 }): Object {
   const {
     result,
   } = diffSelectionSetAgainstStore({
+    context,
     selectionSet,
     rootId,
-    store,
     throwOnMissingField: !returnPartialData,
     variables,
-    fragmentMap,
   });
 
   return result;